Create button.html
In this example, we have a card container (<div class="card-container">) that initially contains three cards. Below the card container, there's a "Load More Cards" button (<button id="load-more-button">) that, when clicked, adds a new card to the container.

The JavaScript code at the end of the HTML document listens for a click event on the "Load More Cards" button. When clicked, it creates a new card element, configures its class and content, and appends it to the card container. You can customize the appearance and content of the cards to suit your needs.